source/transcendence/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bash==0.6
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
msgpack==1.1.0
pillow==10.4.0
priority==1.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
pyotp==2.9.0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
service-identity==24.2.0
setuptools==72.1.0
sqlparse==0.5.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.44.0
zope.interface==7.2
